Re-orgarnized the CDiameterPeer for better usage outside SER Upgraded to support pkg/shm_sums() for memory leak detection Fixed various logging levels

+++ ser_ims/trunk/modules/cdp/globals.h
@@ -67,7 +67,7 @@
 
 #ifdef CDP_FOR_SER
 #else
-extern unsigned int shm_mem_size;
+extern unsigned long shm_mem_size;
 extern int memlog;
 #endif
